+++ Runtime/Scripts/GlobalTools/Utilities/AutoDisableAudioListeners.cs
@@ -5,7 +5,7 @@ namespace PolytopeSolutions.Toolset.GlobalTools.Utilities {
     [RequireComponent(typeof(AudioListener))]
     public class AutoDisableAudioListeners : MonoBehaviour {
         private void Start() {
-            AudioListener[] listeners = GameObject.FindObjectsByType<AudioListener>(FindObjectsSortMode.None);
+            AudioListener[] listeners = GameObject.FindObjectsByType<AudioListener>();
             foreach (AudioListener listener in listeners) {
                 if (listener.gameObject != gameObject)
                     listener.enabled = false;